services/ai-scouting/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.2
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
redis==5.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles==23.2.1
asyncpg==0.29.0
cachetools==5.3.2
tenacity==8.2.3
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2